requirements.txt
Appium-Python-Client==1.2.0
requests==2.26.0
selenium==3.141.0
Pillow==8.3.2